+++ src/cmd/library_migrate.js
@@ -1,6 +1,6 @@
 import {Command, CommandSite} from './command';
 
-import {FileSystemLibraryRepository, FileSystemNamingStrategy} from 'particle-cli-library-manager';
+import {FileSystemLibraryRepository, FileSystemNamingStrategy} from 'particle-library-manager';
 import path from 'path';
 import when from 'when';
 import pipeline from 'when/pipeline';